+++ commands/proposal_submission_new_transfer_test.go
@@ -65,6 +65,7 @@ func testInvalidDestForMetric(t *testing.T) {
 			types.DispatchMetric_DISPATCH_METRIC_RELATIVE_RETURN,
 			types.DispatchMetric_DISPATCH_METRIC_RETURN_VOLATILITY,
 			types.DispatchMetric_DISPATCH_METRIC_VALIDATOR_RANKING,
+			types.DispatchMetric_DISPATCH_METRIC_REALISED_RETURN,
 		},
 		types.AccountType_ACCOUNT_TYPE_REWARD_MAKER_PAID_FEES: {
 			types.DispatchMetric_DISPATCH_METRIC_LP_FEES_RECEIVED,
@@ -74,6 +75,7 @@ func testInvalidDestForMetric(t *testing.T) {
 			types.DispatchMetric_DISPATCH_METRIC_RELATIVE_RETURN,
 			types.DispatchMetric_DISPATCH_METRIC_RETURN_VOLATILITY,
 			types.DispatchMetric_DISPATCH_METRIC_VALIDATOR_RANKING,
+			types.DispatchMetric_DISPATCH_METRIC_REALISED_RETURN,
 		},
 		types.AccountType_ACCOUNT_TYPE_REWARD_MAKER_RECEIVED_FEES: {
 			types.DispatchMetric_DISPATCH_METRIC_LP_FEES_RECEIVED,
@@ -83,6 +85,7 @@ func testInvalidDestForMetric(t *testing.T) {
 			types.DispatchMetric_DISPATCH_METRIC_RELATIVE_RETURN,
 			types.DispatchMetric_DISPATCH_METRIC_RETURN_VOLATILITY,
 			types.DispatchMetric_DISPATCH_METRIC_VALIDATOR_RANKING,
+			types.DispatchMetric_DISPATCH_METRIC_REALISED_RETURN,
 		},
 		types.AccountType_ACCOUNT_TYPE_REWARD_MARKET_PROPOSERS: {
 			types.DispatchMetric_DISPATCH_METRIC_LP_FEES_RECEIVED,
@@ -92,6 +95,7 @@ func testInvalidDestForMetric(t *testing.T) {
 			types.DispatchMetric_DISPATCH_METRIC_RELATIVE_RETURN,
 			types.DispatchMetric_DISPATCH_METRIC_RETURN_VOLATILITY,
 			types.DispatchMetric_DISPATCH_METRIC_VALIDATOR_RANKING,
+			types.DispatchMetric_DISPATCH_METRIC_REALISED_RETURN,
 		},
 		types.AccountType_ACCOUNT_TYPE_REWARD_AVERAGE_POSITION: {
 			types.DispatchMetric_DISPATCH_METRIC_LP_FEES_RECEIVED,
@@ -101,6 +105,7 @@ func testInvalidDestForMetric(t *testing.T) {
 			types.DispatchMetric_DISPATCH_METRIC_RELATIVE_RETURN,
 			types.DispatchMetric_DISPATCH_METRIC_RETURN_VOLATILITY,
 			types.DispatchMetric_DISPATCH_METRIC_VALIDATOR_RANKING,
+			types.DispatchMetric_DISPATCH_METRIC_REALISED_RETURN,
 		},
 		types.AccountType_ACCOUNT_TYPE_REWARD_RELATIVE_RETURN: {
 			types.DispatchMetric_DISPATCH_METRIC_LP_FEES_RECEIVED,
@@ -110,6 +115,7 @@ func testInvalidDestForMetric(t *testing.T) {
 			types.DispatchMetric_DISPATCH_METRIC_AVERAGE_POSITION,
 			types.DispatchMetric_DISPATCH_METRIC_RETURN_VOLATILITY,
 			types.DispatchMetric_DISPATCH_METRIC_VALIDATOR_RANKING,
+			types.DispatchMetric_DISPATCH_METRIC_REALISED_RETURN,
 		},
 		types.AccountType_ACCOUNT_TYPE_REWARD_RETURN_VOLATILITY: {
 			types.DispatchMetric_DISPATCH_METRIC_LP_FEES_RECEIVED,
@@ -119,6 +125,7 @@ func testInvalidDestForMetric(t *testing.T) {
 			types.DispatchMetric_DISPATCH_METRIC_AVERAGE_POSITION,
 			types.DispatchMetric_DISPATCH_METRIC_RELATIVE_RETURN,
 			types.DispatchMetric_DISPATCH_METRIC_VALIDATOR_RANKING,
+			types.DispatchMetric_DISPATCH_METRIC_REALISED_RETURN,
 		},
 		types.AccountType_ACCOUNT_TYPE_REWARD_VALIDATOR_RANKING: {
 			types.DispatchMetric_DISPATCH_METRIC_LP_FEES_RECEIVED,
@@ -128,6 +135,17 @@ func testInvalidDestForMetric(t *testing.T) {
 			types.DispatchMetric_DISPATCH_METRIC_AVERAGE_POSITION,
 			types.DispatchMetric_DISPATCH_METRIC_RELATIVE_RETURN,
 			types.DispatchMetric_DISPATCH_METRIC_RETURN_VOLATILITY,
+			types.DispatchMetric_DISPATCH_METRIC_REALISED_RETURN,
+		},
+		types.AccountType_ACCOUNT_TYPE_REWARD_REALISED_RETURN: {
+			types.DispatchMetric_DISPATCH_METRIC_LP_FEES_RECEIVED,
+			types.DispatchMetric_DISPATCH_METRIC_MAKER_FEES_PAID,
+			types.DispatchMetric_DISPATCH_METRIC_MAKER_FEES_RECEIVED,
+			types.DispatchMetric_DISPATCH_METRIC_MARKET_VALUE,
+			types.DispatchMetric_DISPATCH_METRIC_AVERAGE_POSITION,
+			types.DispatchMetric_DISPATCH_METRIC_RELATIVE_RETURN,
+			types.DispatchMetric_DISPATCH_METRIC_RETURN_VOLATILITY,
+			types.DispatchMetric_DISPATCH_METRIC_VALIDATOR_RANKING,
 		},
 	}
 
@@ -171,6 +189,7 @@ func testInvalidAssetForMetric(t *testing.T) {
 		types.AccountType_ACCOUNT_TYPE_REWARD_AVERAGE_POSITION,
 		types.AccountType_ACCOUNT_TYPE_REWARD_RELATIVE_RETURN,
 		types.AccountType_ACCOUNT_TYPE_REWARD_RETURN_VOLATILITY,
+		types.AccountType_ACCOUNT_TYPE_REWARD_REALISED_RETURN,
 	}
 
 	for _, inv := range invalidTypes {
@@ -242,6 +261,7 @@ func testRecurringWithDispatchInvalidTypes(t *testing.T) {
 	delete(invalidTypes, types.AccountType_ACCOUNT_TYPE_REWARD_VALIDATOR_RANKING)
 	delete(invalidTypes, types.AccountType_ACCOUNT_TYPE_REWARD_RETURN_VOLATILITY)
 	delete(invalidTypes, types.AccountType_ACCOUNT_TYPE_REWARD_RELATIVE_RETURN)
+	delete(invalidTypes, types.AccountType_ACCOUNT_TYPE_REWARD_REALISED_RETURN)
 	delete(invalidTypes, types.AccountType_ACCOUNT_TYPE_REWARD_AVERAGE_POSITION)
 	delete(invalidTypes, types.AccountType_ACCOUNT_TYPE_INSURANCE)
 	delete(invalidTypes, types.AccountType_ACCOUNT_TYPE_GENERAL)
@@ -580,7 +600,7 @@ func testNewTransferChangeSubmissionInvalidDestinationTypeFails(t *testing.T) {
 	delete(allAccountTypes, int32(types.AccountType_ACCOUNT_TYPE_REWARD_RELATIVE_RETURN))
 	delete(allAccountTypes, int32(types.AccountType_ACCOUNT_TYPE_REWARD_RETURN_VOLATILITY))
 	delete(allAccountTypes, int32(types.AccountType_ACCOUNT_TYPE_REWARD_VALIDATOR_RANKING))
-
+	delete(allAccountTypes, int32(types.AccountType_ACCOUNT_TYPE_REWARD_REALISED_RETURN))
 	delete(allAccountTypes, int32(types.AccountType_ACCOUNT_TYPE_UNSPECIFIED))
 
 	for at := range allAccountTypes {